requirements.txt
torchvision
IPython
transformers
scipy
opencv-python
torchmetrics
matplotlib
tqdm